requirements.txt
beautifulsoup4==4.6.0
elasticsearch==6.2.0
feedparser==5.2.1
HTMLParser==0.0.2
lxml==4.2.1
numpy==1.14.2
pandas==0.22.0
python-dateutil==2.7.2
pytz==2018.3
six==1.11.0
urllib3==1.22
